requirements.txt
Django
django-crispy-forms
django_import_export
django-waitress
PyMySQL
waitress